started working with redux

+++ src/firebase.js
@@ -2,6 +2,7 @@ import { initializeApp } from "firebase/app"
 import { getStorage, ref, uploadBytesResumable, getDownloadURL, deleteObject } from "firebase/storage"
 import {getFirestore} from "@firebase/firestore"
 import { collection, addDoc, onSnapshot, serverTimestamp, query, orderBy, deleteDoc, doc  } from "firebase/firestore"; 
+import { getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword, updateProfile, onAuthStateChanged, signOut, GoogleAuthProvider, signInWithPopup  } from "firebase/auth";
 
 const firebaseConfig = {
   apiKey: "AIzaSyCIc4XWBbUFO0Tkdo2a6dvATPvo7-S2y4A",
@@ -16,4 +17,7 @@ const firebaseConfig = {
 const app = initializeApp(firebaseConfig)
 const storage = getStorage(app)
 const db = getFirestore(app)
-export {collection, addDoc, db, onSnapshot, serverTimestamp,query,orderBy, storage,ref,uploadBytesResumable,getDownloadURL,deleteDoc, doc, deleteObject}
+const auth = getAuth(app);
+const googleProvider = new GoogleAuthProvider()
+
+export {googleProvider, collection, addDoc, db, onSnapshot, serverTimestamp,query,orderBy, storage,ref,uploadBytesResumable,getDownloadURL,deleteDoc, doc, deleteObject, auth, createUserWithEmailAndPassword, signInWithEmailAndPassword, updateProfile, onAuthStateChanged, signOut, signInWithPopup }
